+++ config.ini
@@ -11,7 +11,7 @@ supported_lang_codes = en,bn,gu,hi,kn,ml,mr,or,pa,ta,te
 support_response_format = text,audio
 
 [llm]
-gpt_model=gpt-4
+gpt_model=myjp_gpt4
 enable_bot_intent=false
 intent_prompt = Identify if the user's query is about the bot's persona or 'Teacher Tara' or 'Parent Tara'. If yes, return the answer as 'Yes' else return answer as 'No' only.
 bot_prompt = {
