+++ main.py
@@ -9,7 +9,7 @@ from pydantic import BaseModel
 from storage.api import *
 from io_processing import *
 # from query_with_langchain import *
-from llm.api import querying_with_langchain_gpt3
+from query_with_langchain import *
 from telemetry_middleware import TelemetryMiddleware
 from config_util import get_config_value
 from utils import *
